requirements.txt
tensorflow-gpu==2.3.0
numpy==1.18.5
